requirements.txt
requests
python-i18n[YAML]
python-i18n